# [msbuild] Sort inputs when doing aot compilation and native compilation.