Sort inputs when doing AOT compilation and native code compilation, so that we
start by processing the input that takes the longest to process. This helps
maximizing usage of parallel resources, because there's a smaller chance we'll
be left at the end working on a single long-running task.

Some tentative timings from compiling monotouch-test locally, it shows a few
seconds improvement (which percentage-wise can be significant).

| Task                    | Before | After |
| ----------------------- | ------ | ----- |
| Native code compilation | ~25s   |  ~22s |
| AOT compilation         | ~23s   |  ~22s |