@vs-mobiletools-engineering-service2 2025-05-07
# :white_check_mark: [CI Build #95ce31e] Build passed (Build macOS tests) :white_check_mark:




[Pipeline](https://dev.azure.com/devdiv/DevDiv/_build/index?buildId=11550800&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [95ce31e49c6ca6b43f36b49e801d88320ce4531f](https://github.com/dotnet/macios/pull/22755/commits/95ce31e49c6ca6b43f36b49e801d88320ce4531f)  [PR build]

[comment]: <> (This is a comment added by Azure DevOps, id: build Build macOS tests)