@vs-mobiletools-engineering-service2 2025-05-07
# :computer: [CI Build #95ce31e] Tests on macOS X64 - Mac Sonoma (14) passed :computer:

:white_check_mark: **All** tests on macOS X64 - Mac Sonoma (14) passed.

[Pipeline](https://dev.azure.com/devdiv/DevDiv/_build/index?buildId=11550800&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [95ce31e49c6ca6b43f36b49e801d88320ce4531f](https://github.com/dotnet/macios/pull/22755/commits/95ce31e49c6ca6b43f36b49e801d88320ce4531f)  [PR build]

[comment]: <> (This is a comment added by Azure DevOps, id: macOS X64 - Mac Sonoma _14_)